requirements.txt
astroid==1.6.4
certifi==2018.4.16
chardet==3.0.4
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.1
python-dateutil==2.7.3
python-dotenv==0.8.2
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
